yoyo-db-migration/requirements.txt
certifi==2019.11.28
iniherit==0.3.9
PyMySQL==0.9.3
sqlparse==0.3.0
tabulate==0.8.9
text-unidecode==1.3
yoyo-migrations==7.3.2